Verifier: extract Bytecode related structs into separate module; update BytecodeMismatch error
