+++ smart-contract-verifier/src/solidity/mod.rs
@@ -1,6 +1,7 @@
 mod errors;
 mod verifier;
 
+mod bytecode;
 mod metadata;
 
 pub(crate) use verifier::{VerificationSuccess, Verifier};
