+++ smart-contract-verifier/src/solidity/verifier.rs
@@ -1,285 +1,15 @@
-use super::errors::{BytecodeInitializationError, VerificationError};
+use super::{
+    bytecode::{Bytecode, BytecodePart, LocalBytecode},
+    errors::{BytecodeInitializationError, VerificationError},
+};
 use crate::{
     solidity::{errors::VerificationErrorKind, metadata::MetadataHash},
     types::Mismatch,
     DisplayBytes,
 };
-use bytes::{Buf, Bytes};
+use bytes::Bytes;
 use ethabi::{Constructor, Token};
 use ethers_solc::{artifacts::Contract, Artifact, CompilerOutput};
-use std::str::FromStr;
-
-/// Combine creation_tx_input and deployed_bytecode.
-/// Guarantees that `deployed_bytecode` was actually deployed
-/// by `creation_tx_input`.
-#[derive(Clone, Debug, PartialEq, Eq)]
-struct Bytecode {
-    /// Raw bytecode bytes used in contract creation transaction
-    creation_tx_input: Bytes,
-    /// Raw deployed bytecode bytes
-    deployed_bytecode: Bytes,
-
-    /// Hex representation of creation tx input without "0x" prefix
-    creation_tx_input_str: String,
-    /// Hex representation of deployed bytecode without "0x" prefix
-    deployed_bytecode_str: String,
-}
-
-impl Bytecode {
-    pub fn new(
-        creation_tx_input: &str,
-        deployed_bytecode: &str,
-    ) -> Result<Self, BytecodeInitializationError> {
-        let creation_tx_input = DisplayBytes::from_str(creation_tx_input)
-            .map_err(|_| {
-                BytecodeInitializationError::InvalidCreationTxInput(creation_tx_input.to_string())
-            })?
-            .0;
-        if creation_tx_input.is_empty() {
-            return Err(BytecodeInitializationError::EmptyCreationTxInput);
-        }
-
-        let deployed_bytecode = DisplayBytes::from_str(deployed_bytecode)
-            .map_err(|_| {
-                BytecodeInitializationError::InvalidDeployedBytecode(deployed_bytecode.to_string())
-            })?
-            .0;
-        if deployed_bytecode.is_empty() {
-            return Err(BytecodeInitializationError::EmptyDeployedBytecode);
-        }
-
-        // We transform `&str` into `Bytes` initially, so that we can ensure that initial input is a valid hex string
-        // and enforce internal `Bytecode` representation not to have `0x` prefix in it.
-        Self::from_bytes(creation_tx_input, deployed_bytecode)
-    }
-
-    pub fn from_bytes(
-        creation_tx_input: Bytes,
-        deployed_bytecode: Bytes,
-    ) -> Result<Self, BytecodeInitializationError> {
-        // `hex::encode` encodes bytes as hex string without "0x" prefix
-        let creation_tx_input_str = hex::encode(&creation_tx_input);
-        let deployed_bytecode_str = hex::encode(&deployed_bytecode);
-
-        Ok(Self {
-            creation_tx_input,
-            deployed_bytecode,
-
-            creation_tx_input_str,
-            deployed_bytecode_str,
-        })
-    }
-}
-
-impl TryFrom<&Contract> for Bytecode {
-    type Error = BytecodeInitializationError;
-
-    fn try_from(contract: &Contract) -> Result<Self, Self::Error> {
-        let deployed_bytecode = {
-            contract.get_deployed_bytecode_bytes().ok_or_else(|| {
-                let bytecode = contract
-                    .get_deployed_bytecode_object()
-                    .unwrap_or_default()
-                    .as_str()
-                    .unwrap_or_default()
-                    .to_string();
-                BytecodeInitializationError::InvalidDeployedBytecode(bytecode)
-            })?
-        };
-        let creation_tx_input = {
-            contract.get_bytecode_bytes().ok_or_else(|| {
-                let bytecode = contract
-                    .get_bytecode_object()
-                    .unwrap_or_default()
-                    .as_str()
-                    .unwrap_or_default()
-                    .to_string();
-                BytecodeInitializationError::InvalidCreationTxInput(bytecode)
-            })?
-        };
-        Bytecode::from_bytes(creation_tx_input.0.clone(), deployed_bytecode.0.clone())
-    }
-}
-
-#[derive(Clone, Debug, PartialEq, Eq)]
-enum BytecodePart {
-    Main {
-        raw: Bytes,
-    },
-    Metadata {
-        metadata_raw: Bytes,
-        metadata: MetadataHash,
-        metadata_length_raw: Bytes,
-    },
-}
-
-impl BytecodePart {
-    pub fn size(&self) -> usize {
-        match self {
-            BytecodePart::Main { raw } => raw.len(),
-            BytecodePart::Metadata {
-                metadata_raw,
-                metadata_length_raw,
-                ..
-            } => metadata_raw.len() + metadata_length_raw.len(),
-        }
-    }
-}
-
-#[derive(Clone, Debug, PartialEq, Eq)]
-struct RemoteBytecode {
-    bytecode: Bytecode,
-}
-
-impl RemoteBytecode {
-    pub fn new(bytecode: Bytecode) -> Self {
-        Self { bytecode }
-    }
-
-    pub fn creation_tx_input(&self) -> &Bytes {
-        &self.bytecode.creation_tx_input
-    }
-}
-
-/// Encapsulates result of local source code compilation.
-/// Splits compiled creation transaction input.
-#[derive(Clone, Debug, PartialEq, Eq)]
-struct LocalBytecode {
-    bytecode: Bytecode,
-    creation_tx_input_parts: Vec<BytecodePart>,
-}
-
-impl LocalBytecode {
-    /// Initializes a new [`LocalBytecode`] struct.
-    ///
-    /// `bytecode` and `bytecode_modified` must differ only in metadata hashes.
-    ///
-    /// Any error here is [`VerificationErrorKind::InternalError`], as both original
-    /// and modified bytecodes are obtained as a result of local compilation.
-    pub fn new(
-        bytecode: Bytecode,
-        bytecode_modified: Bytecode,
-    ) -> Result<Self, VerificationErrorKind> {
-        let creation_tx_input_parts = Self::split(
-            &bytecode.creation_tx_input,
-            &bytecode_modified.creation_tx_input,
-        )?;
-
-        Ok(Self {
-            bytecode,
-            creation_tx_input_parts,
-        })
-    }
-
-    pub fn creation_tx_input(&self) -> &Bytes {
-        &self.bytecode.creation_tx_input
-    }
-
-    /// Splits bytecode onto [`BytecodePart`]s using bytecode with modified metadata hashes.
-    ///
-    /// Any error here is [`VerificationErrorKind::InternalError`], as both original
-    /// and modified bytecodes are obtained as a result of local compilation.
-    fn split(
-        raw: &Bytes,
-        raw_modified: &Bytes,
-    ) -> Result<Vec<BytecodePart>, VerificationErrorKind> {
-        if raw.len() != raw_modified.len() {
-            return Err(VerificationErrorKind::InternalError(format!(
-                "bytecode and modified bytecode length mismatch: {}",
-                Mismatch::new(raw.len(), raw_modified.len())
-            )));
-        }
-
-        let parts_total_size = |parts: &Vec<BytecodePart>| -> usize {
-            parts.iter().fold(0, |size, el| size + el.size())
-        };
-
-        let mut result = Vec::new();
-
-        let mut i = 0usize;
-        while i < raw.len() {
-            let decoded = Self::parse_bytecode_parts(&raw.slice(i..raw.len()), &raw_modified[i..])?;
-            let decoded_size = parts_total_size(&decoded);
-            result.extend(decoded);
-
-            i += decoded_size;
-        }
-
-        Ok(result)
-    }
-
-    /// Finds the next [`BytecodePart`]s into a series of bytes.
-    ///
-    /// Parses at most one [`BytecodePart::Main`] and one [`BytecodePart::Metadata`].
-    fn parse_bytecode_parts(
-        raw: &Bytes,
-        raw_modified: &[u8],
-    ) -> Result<Vec<BytecodePart>, VerificationErrorKind> {
-        let mut parts = Vec::new();
-
-        let len = raw.len();
-
-        let mut i = 0usize;
-        while i < len {
-            if raw[i] == raw_modified[i] {
-                i += 1;
-                continue;
-            }
-
-            // The first different byte. The metadata hash itself started somewhere earlier
-            // (at least for "a1"/"a2" indicating number of elements in cbor mapping).
-            // Next steps are trying to find that beginning.
-
-            let mut result = MetadataHash::from_cbor(&raw[i..]);
-            while result.is_err() {
-                // It is the beginning of the bytecode segment but no metadata hash has been parsed
-                if i == 0 {
-                    return Err(VerificationErrorKind::InternalError(
-                        "failed to parse bytecode part".into(),
-                    ));
-                }
-                i -= 1;
-
-                result = MetadataHash::from_cbor(&raw[i..]);
-            }
-
-            let (metadata, metadata_length) = result.unwrap();
-
-            if len < i + metadata_length + 2 {
-                return Err(VerificationErrorKind::InternalError(
-                    "failed to parse metadata length".into(),
-                ));
-            }
-            // Decode length of metadata hash representation
-            let metadata_length_raw = raw.slice((i + metadata_length)..(i + metadata_length + 2));
-            let encoded_metadata_length = metadata_length_raw.clone().get_u16() as usize;
-            if encoded_metadata_length != metadata_length {
-                return Err(VerificationErrorKind::InternalError(format!(
-                    "encoded metadata length does not correspond to actual metadata length: {}",
-                    Mismatch::new(metadata_length, encoded_metadata_length)
-                )));
-            }
-
-            parts.push(BytecodePart::Metadata {
-                metadata_raw: raw.slice(i..(i + metadata_length)),
-                metadata,
-                metadata_length_raw,
-            });
-            break;
-        }
-
-        if i > 0 {
-            parts.insert(
-                0,
-                BytecodePart::Main {
-                    raw: raw.slice(0..i),
-                },
-            )
-        }
-
-        Ok(parts)
-    }
-}
 
 /// Verifier used for contract verification.
 ///
@@ -287,7 +17,7 @@ impl LocalBytecode {
 /// further be used in verification process.
 #[derive(Clone, Debug, PartialEq, Eq)]
 pub struct Verifier {
-    remote_bytecode: RemoteBytecode,
+    remote_bytecode: Bytecode,
 }
 
 /// The structure returned as a result when verification successes.
@@ -311,7 +41,7 @@ impl Verifier {
     ) -> Result<Self, BytecodeInitializationError> {
         let bytecode = Bytecode::new(creation_tx_input, deployed_bytecode)?;
         Ok(Self {
-            remote_bytecode: RemoteBytecode::new(bytecode),
+            remote_bytecode: bytecode,
         })
     }
 
@@ -395,6 +125,7 @@ impl Verifier {
         Err(errors)
     }
 
+    /// Tries to verify the remote bytecode via locally compiled contract.
     fn compare(
         &self,
         contract: &Contract,
@@ -434,28 +165,40 @@ impl Verifier {
     }
 
     fn compare_creation_tx_inputs(
-        remote_bytecode: &RemoteBytecode,
+        remote_bytecode: &Bytecode,
         local_bytecode: &LocalBytecode,
     ) -> Result<(), VerificationErrorKind> {
         let remote_creation_tx_input = remote_bytecode.creation_tx_input();
         let local_creation_tx_input = local_bytecode.creation_tx_input();
 
         if remote_creation_tx_input.len() < local_creation_tx_input.len() {
-            return Err(VerificationErrorKind::BytecodeMismatch(Mismatch::new(
-                local_creation_tx_input.clone().into(),
-                remote_creation_tx_input.clone().into(),
-            )));
+            return Err(VerificationErrorKind::BytecodeLengthMismatch {
+                part: Mismatch::new(
+                    local_creation_tx_input.len(),
+                    remote_creation_tx_input.len(),
+                ),
+                raw: Mismatch::new(
+                    local_creation_tx_input.clone().into(),
+                    remote_creation_tx_input.clone().into(),
+                ),
+            });
         }
 
         Self::compare_bytecode_parts(
             remote_creation_tx_input,
             local_creation_tx_input,
-            &local_bytecode.creation_tx_input_parts,
+            local_bytecode.creation_tx_input_parts(),
         )?;
 
         Ok(())
     }
 
+    /// Performs an actual comparison of locally compiled bytecode
+    /// with remote bytecode provided for verification.
+    ///
+    /// # Panics
+    ///
+    /// The function will panic if `remote_raw.len()` is less than `local_raw.len()`.
     fn compare_bytecode_parts(
         remote_raw: &Bytes,
         local_raw: &Bytes,
@@ -473,10 +216,13 @@ impl Verifier {
             match part {
                 BytecodePart::Main { raw } => {
                     if raw != &remote_raw[i..i + raw.len()] {
-                        return Err(VerificationErrorKind::BytecodeMismatch(Mismatch::new(
-                            local_raw.clone().into(),
-                            remote_raw.clone().into(),
-                        )));
+                        return Err(VerificationErrorKind::BytecodeMismatch {
+                            part: Mismatch::new(
+                                raw.clone().into(),
+                                remote_raw.slice(i..i + raw.len()).into(),
+                            ),
+                            raw: Mismatch::new(local_raw.clone().into(), remote_raw.clone().into()),
+                        });
                     }
                 }
                 BytecodePart::Metadata {
@@ -484,11 +230,11 @@ impl Verifier {
                     metadata_length_raw,
                     ..
                 } => {
-                    let (remote_metadata, remote_metadata_size) =
+                    let (remote_metadata, remote_metadata_length) =
                         MetadataHash::from_cbor(&remote_raw[i..])
                             .map_err(|err| VerificationErrorKind::MetadataParse(err.to_string()))?;
 
-                    let start_index = i + remote_metadata_size;
+                    let start_index = i + remote_metadata_length;
                     if &remote_raw[start_index..start_index + 2] != metadata_length_raw {
                         return Err(VerificationErrorKind::MetadataParse(
                             "metadata length mismatch".into(),
@@ -519,9 +265,6 @@ impl Verifier {
 
     /// Extracts constructor arguments from the creation transaction input specified on
     /// [`Verifier`] initialization.
-    ///
-    /// Returns `Err` if constructor arguments cannot be extracted (should not be the case
-    /// if `Bytecode.verify_bytecode_with_extra_data` was called before).
     fn extract_constructor_args(
         remote_raw: &Bytes,
         local_raw: &Bytes,
@@ -559,7 +302,6 @@ impl Verifier {
     ///
     /// Returns `Err` if bytes do not correspond to the constructor arguments representation.
     fn parse_constructor_args(
-        // &self,
         encoded_args: Bytes,
         abi_constructor: &Constructor,
     ) -> Result<Vec<Token>, VerificationErrorKind> {
